Split out event source and event flow for VAI
