+++ provider-service/base/pkg/streams/sinks/error_sink.go
