+++ provider-service/kfp/cmd/main.go
@@ -7,8 +7,8 @@ import (
 	"github.com/sky-uk/kfp-operator/argo/common"
 	. "github.com/sky-uk/kfp-operator/provider-service/base/pkg"
 	configLoader "github.com/sky-uk/kfp-operator/provider-service/base/pkg/config"
-	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/sinks"
-	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/sources"
+	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/streams/sinks"
+	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/streams/sources"
 	kfp "github.com/sky-uk/kfp-operator/provider-service/kfp/internal"
 	"go.uber.org/zap/zapcore"
 	"os"
