+++ provider-service/kfp/internal/suite_decoupled_test.go
@@ -22,8 +22,8 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/sky-uk/kfp-operator/argo/common"
 	"github.com/sky-uk/kfp-operator/provider-service/base/pkg"
-	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/sinks"
-	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/sources"
+	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/streams/sinks"
+	"github.com/sky-uk/kfp-operator/provider-service/base/pkg/streams/sources"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/util/rand"
