+++ provider-service/vai/internal/vai_event_flow_unit_test.go
@@ -62,6 +62,7 @@ var _ = Context("VaiEventingServer", func() {
 		Entry("Paused", aiplatformpb.PipelineState_PIPELINE_STATE_PAUSED),
 	)
 
+	//flow
 	DescribeTable("toRunCompletionEventData for job that has completed", func(pipelineState aiplatformpb.PipelineState, status common.RunCompletionStatus) {
 		runConfigurationName := common.RandomNamespacedName()
 		pipelineName := common.RandomNamespacedName()
@@ -131,6 +132,7 @@ var _ = Context("VaiEventingServer", func() {
 		Entry("Pending", aiplatformpb.PipelineState_PIPELINE_STATE_CANCELLED, common.RunCompletionStatuses.Failed),
 	)
 
+	//flow
 	Describe("artifactsFilterData", func() {
 		When("The job is missing the component", func() {
 			It("Produces no artifacts", func() {
@@ -195,6 +197,7 @@ var _ = Context("VaiEventingServer", func() {
 		})
 	})
 
+	//flow
 	Describe("Legacy: modelServingArtifactsForJob", func() {
 		When("The job has an output with an artifact that doesn't match the SchemaTitle", func() {
 			It("Produces no servingModelArtifacts", func() {
@@ -422,6 +425,7 @@ var _ = Context("VaiEventingServer", func() {
 		})
 	})
 
+	//flow
 	Describe("runCompletionEventDataForRun", func() {
 		When("GetPipelineJob errors", func() {
 			It("returns no event", func() {
